udp: Winsock datagram I/O (udp_io_win) — build UDP on Windows
Step 3 of the UDP Windows port: implement the datagram-I/O seam for
Winsock and wire the whole UDP subsystem into the Windows build. With
udp_io_posix.c's platform-neutral seam in place (step 1), this adds the
Windows sibling and makes the portable layer (udp.c / udp_server.c)
actually compile on Winsock. `make OS=windows` now builds and links the
entire library except the DNS resolver — TCP HTTP core + UDP datagram
core in one Windows PE.

  - src/udp_io_win.c (new): the Winsock datagram I/O impl, mirroring
    udp_io_posix.c. WSARecvMsg/WSASendMsg are fetched once per process via
    WSAIoctl(SIO_GET_EXTENSION_FUNCTION_POINTER, WSAID_*) and cached; cmsg
    build/parse uses WSA_CMSG_* with IN_PKTINFO/IN6_PKTINFO (local-address
    recovery) and IP_TOS/IPV6_TCLASS. No recvmmsg/GSO on Windows, so the
    batch inits are no-ops and send_gso returns -1/EIO (non-would-block) so
    udp.c rules GSO unsupported and falls back to per-segment sends. An
    errno shim (WSAEWOULDBLOCK -> EWOULDBLOCK, else EIO) lets udp.c's errno
    checks work unchanged. Keeps its intra-Winsock IP_*/IPV6_* #ifdefs
    here (allowed in a platform TU, like udp_io_posix.c).

  - src/udp.c: portability fixes so the portable layer builds on Winsock,
    no platform #ifdef added. Dropped four redundant POSIX net includes
    (netdb/arpa-inet/netinet-in/sys-socket — MinGW lacks them; the seam's
    socket.h -> sockcompat.h already covers sockaddr/getaddrinfo/inet_ntop/
    ntohs). Cast the four socklen_t-vs-sizeof comparisons to size_t (Windows
    socklen_t is signed int -> sign-compare under -Werror). Also routed the
    last raw getsockname (kl_udp_local_port) through the seam
    (kl_sock_get_local_addr), the same sweep the TCP TUs already had.

  - src/udp_server.c: cast the one socklen_t-vs-sizeof comparison to size_t.

  - Makefile: UDP_IO_SRC = src/udp_io_win.c in the Windows branch; the
    Windows filter-out now drops only dns_resolver.c (UDP is built). New
    smoke-udp target (tests/smoke_udp) — a single-threaded event-loop
    loopback datagram roundtrip with recv_pktinfo on, so the WSARecvMsg/
    cmsg path runs (not just recvfrom). No -lpthread (unlike smoke-tcp).

  - tests/smoke_udp.c (new): the datagram link + roundtrip smoke test.

  - .github/workflows/ci.yml: the windows-latest (MinGW) job now builds
    TCP + UDP and runs both smoke-tcp and smoke-udp — the authoritative
    runtime gate for udp_io_win.c on real Windows.

Verified: macOS full suite + smoke-tcp/smoke-udp (roundtrip OK,
local_addr recovered — the cmsg path exercised); Linux glibc full suite;
debug-test (ASan+UBSan); scan-build (no bugs); make cppcheck (51 files
clean); musl build+test; MinGW full Windows build + smoke_tcp.exe /
smoke_udp.exe link (UDP included). Windows runtime is gated by the
windows-latest CI job (local amd64-Wine-under-Rosetta is too flaky to
trust).

+++ src/udp_server.c
@@ -10,7 +10,7 @@ static void udp_server_on_recv(KlUdp *udp, const void *data, size_t len,
                                void *user_data) {
     (void)udp;
     KlUdpServer *s = user_data;
-    if (local && local_len && local_len <= sizeof(s->local)) {
+    if (local && local_len && (size_t)local_len <= sizeof(s->local)) {
         memcpy(&s->local, local, local_len);
         s->local_len = local_len;
     } else {
